Update base image to allow UID and GID to be set on container startup
To achieve this, I have switched out our init to always run as root for all containers. This will carry out any root user config and then, if the container should exec the process as an unprivileged user, this init will use gosu to drop down to that user.

This work removes all requirement for sudo inside the container.
This has the benefit of PID 1 always being the service that is running rather than sudo exec command. PID 1 will always receive all SIG calls from Docker and therefore, when we terminate the container we will be making that cll directly to the service running inside - EG. sunshine, steam, etc.