Configure default user for services that will run as root
X, pulse and udev all need to run as the root user (not retro). Reset the 'UNAME' and 'HOME' vars set for these images to "root" and "/root" respectively.
Resetting these variables will modify the init scripts slightly to skip the retro user setup and to execute the run scripts as root with gosu.